Moved the FilePreview implementation to File.container and just export it in index.js
