+++ src/file/index.js
@@ -23,87 +23,8 @@
  *  Module for file rendering in Renga.
  */
 
-import React from 'react';
-import ReactDom from 'react-dom';
-import ReactMarkdown from 'react-markdown'
-import hljs from 'highlight.js'
-import { JupyterNotebook } from './File.container'
-import '../../node_modules/highlight.js/styles/atom-one-light.css'
+
+import { FilePreview } from './File.container'
 
 
-const IMAGE_EXTENSIONS = ['jpg', 'jpeg', 'png', 'tiff', 'pdf', 'gif'];
-const CODE_EXTENSIONS = ['py', 'js', 'json', 'sh', 'r', 'txt',];
-
-
-// FIXME: Unify the file viewing for kus (embedded) and independent file viewing.
-// FIXME: Javascript highlighting is broken for large files.
-// FIXME: Fix positioning of input tags when rendering Jupyter Notebooks.
-
-export class FilePreview extends React.Component {
-
-  getFileExtension = () => {
-    if (!this.props.file) {
-      return null
-    } else {
-      return this.props.file.file_name
-        .match(/\.(.*)/)[0]
-        .slice(1,)
-        .toLowerCase();
-    }
-  };
-
-  fileIsCode = () => CODE_EXTENSIONS.indexOf(this.getFileExtension()) >= 0;
-  fileIsImage = () => IMAGE_EXTENSIONS.indexOf(this.getFileExtension()) >= 0;
-
-  highlightBlock = () => {
-    // FIXME: Usage of findDOMNode is discouraged.
-    // eslint-disable-next-line
-    const baseNode = ReactDom.findDOMNode(this);
-    hljs.highlightBlock(baseNode);
-  };
-
-  componentDidMount(){
-    if (this.fileIsCode()) this.highlightBlock()
-  }
-  componentDidUpdate(){
-    if (this.fileIsCode()) this.highlightBlock()
-  }
-
-  render(){
-    // File has not yet been fetched
-    if (!this.props.file) {
-      return <p>Loading...</p>;
-    }
-    // Various types of images
-    if (this.fileIsImage()) {
-      return <img
-        className="image-preview"
-        alt={this.props.file.file_name}
-        src={'data:image;base64,' + this.props.file.content}
-      />;
-    }
-    // Code with syntax highlighting
-    if (this.fileIsCode()) {
-      return (
-        <pre className={`hljs ${this.getFileExtension()}`}>
-          <code>{atob(this.props.file.content)}</code>
-        </pre>
-      );
-    }
-    // Markdown
-    if (this.getFileExtension() === 'md'){
-      return <ReactMarkdown source={atob(this.props.file.content)}/>;
-    }
-    // Jupyter Notebook
-    if (this.getFileExtension() === 'ipynb'){
-      return <JupyterNotebook
-        notebook={JSON.parse(atob(this.props.file.content))}
-        filePath={this.props.file.file_path}
-        projectId={this.props.projectId}
-        client={this.props.client}/>;
-    }
-
-    // File extension not supported
-    return <p>{`Unable to preview file with extension .${this.getFileExtension()}`}</p>;
-  }
-}
+export { FilePreview }
